requirement_text.txt
streamlit
pandas
numpy
torch
transformers
faiss-cpu
sentence-transformers
